Player position correctly changes with new levels
Currently it's set so that once you have all access cards, when you move next you will transport to the next level. I noticed that if I moved right or left, then the y coordinate the player transported to would be correct, but the x coordinate would be the same as if I had moved one space. So, I figured that it was what I thought yesterday and the position updates in the move function were overriding what was done in load level.

To fix this I made a boolean variable called level_end which is set to false until you collect all access cards and move. Then I just changed the update function so that it only does all of the movement stuff if level_end is false.